// app/types.ts
export type RootStackParamList = {
    Tabs: undefined;
    Details: {
      item: {
        id: string;
        title: string;
        image: any;
        description: string;
        category: string;
        mapUrl: string;
        gallery?: any[];
      };
    };
    About: undefined;
    BookingForm: {
      item: {
        id: string;
        title: string;
        image: any;
        description: string;
        category: string;
        mapUrl: string;
        gallery?: any[];
      };
    };
  };
  